feat: netsim integration tests (#15)

+++ netsim/main.py
@@ -13,8 +13,25 @@ from parser import stats_parser
 from network import StarTopo
 
 TIMEOUT = 60 * 2
+
+def logs_on_error(nodes, prefix):
+    node_counts = {}
+    for node in nodes:
+        node_counts[node['name']] = int(node['count'])
+        for i in range(int(node['count'])):
+            node_name = '%s_%d' % (node['name'], i)
+            log_name= 'logs/%s__%s.txt' % (prefix, node_name)
+            if os.path.isfile(log_name):
+                print('\n\n[INFO] Log file: %s' % log_name)
+                f = open(log_name, 'r')
+                lines = f.readlines()
+                for line in lines:
+                    print('[INFO][%s__%s] %s' % (prefix, node_name, line.rstrip()))
+            else:
+                print('[WARN] log file missing: %s' % log_name)
+    pass
     
-def run(nodes, prefix):
+def run(nodes, prefix, integration, debug=False):
     topo = StarTopo(nodes=nodes)
     net = Mininet(topo = topo, waitConnected=True, link=TCLink)
     net.start()
@@ -23,11 +40,16 @@ def run(nodes, prefix):
     print( "Testing network connectivity" )
     net.pingAll()
 
+    env_vars = os.environ.copy()
+    if debug:
+        env_vars['RUST_LOG'] = 'debug'
+
     p_box = []
     p_short_box = []
 
     node_counts = {}
     node_ips = {}
+    node_params = {}
 
     for node in nodes:
         node_counts[node['name']] = int(node['count'])
@@ -52,13 +74,30 @@ def run(nodes, prefix):
                 connect_to = '%s_%d' % (node['connect']['node'], id)
                 ip = node_ips[connect_to]
                 cmd = cmd % (ip, id)
-            p = n.popen(cmd, stdout=f, stderr=f, shell=True)
+            if node['connect']['strategy'] == 'params':
+                cnt = node_counts[node['connect']['node']]
+                id = i % cnt
+                connect_to = '%s_%d' % (node['connect']['node'], id)
+                param = node_params[connect_to]
+                cmd = cmd % (param)
+            p = n.popen(cmd, stdout=f, stderr=f, shell=True, env=env_vars)
             if 'process' in node and node['process'] == 'short':
                 p_short_box.append(p)
             else:
                 p_box.append(p)
         if 'wait' in node:
             time.sleep(node['wait'])
+        if 'param_parser' in node:
+            for i in range(int(node['count'])):
+                node_name = '%s_%d' % (node['name'], i)
+                n = net.get(node_name)
+                f = open('logs/%s__%s.txt' % (prefix, node_name), 'r')
+                lines = f.readlines()
+                for line in lines:
+                    if node['param_parser'] == 'iroh_ticket':
+                        if line.startswith('All-in-one ticket'):
+                            node_params[node_name] = line[len('All-in-one ticket: '):].strip()
+                            break
 
     # CLI(net)
     for i in range(TIMEOUT):
@@ -66,7 +105,17 @@ def run(nodes, prefix):
         if not any(p.poll() is None for p in p_short_box):
             break
     for p in p_short_box:
-        p.terminate()
+        if integration:
+            r = p.poll()
+            if r is None:
+                p.terminate()
+                logs_on_error(nodes, prefix)
+                raise Exception('Process has timed out')
+            if r != 0:
+                logs_on_error(nodes, prefix)
+                raise Exception('Process has failed')
+        else:
+            p.terminate()
     for p in p_box:
         p.terminate()
     net.stop()
@@ -77,6 +126,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument("cfg", help = "Input config file")
     parser.add_argument("-r", help = "Run only report generation", action='store_true')
+    parser.add_argument("--integration", help = "Run in integration test mode", action='store_true')
     args = parser.parse_args()
 
     paths = []
@@ -100,5 +150,5 @@ if __name__ == '__main__':
             nodes = case['nodes']
             print('running "%s"...' % prefix)
             if not args.r:
-                run(nodes, prefix)
+                run(nodes, prefix, args.integration, True)
             stats_parser(nodes, prefix)
